+++ eWAY.Rapid/Internals/Request/CancelAuthorisationRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace eWAY.Rapid.Internals.Request
-{
-    internal class CancelAuthorisationRequest: CaptureAuthBaseRequest
-    {
-    }
-}