+++ eWAY.Rapid/Internals/Response/CancelAuthorisationResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace eWAY.Rapid.Internals.Response
-{
-    internal class CancelAuthorisationResponse: CaptureAuthBaseResponse
-    {
-    }
-}