+++ eWAY.Rapid/Internals/Response/CapturePaymentResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace eWAY.Rapid.Internals.Response
-{
-    internal class CapturePaymentResponse: CaptureAuthBaseResponse
-    {
-    }
-}