+++ eWAY.Rapid/Internals/Response/DirectCancelAuthorisationResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace eWAY.Rapid.Internals.Response
+{
+    internal class DirectCancelAuthorisationResponse: CaptureAuthBaseResponse
+    {
+    }
+}