+++ eWAY.Rapid/Models/CancelAuthorisationRequest.cs
@@ -1,0 +1,13 @@
+﻿namespace eWAY.Rapid.Models
+{
+    /// <summary>
+    /// The Transaction ID of the Authorisation to capture
+    /// </summary>
+    public class CancelAuthorisationRequest
+    {
+        /// <summary>
+        /// The Transaction ID of the Authorisation to capture
+        /// </summary>
+        public string TransactionId { get; set; }
+    }
+}