+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All Notable changes will be documented in this file
 
+## 1.1.0.0
+
+ - Added Capture & Cancel functions for pre-auth payments
+ - Added Email field for customers
+ - Marked Customer.CustomerIP as obselete, Transaction.CustomerIP should be used
+
 ## 1.0.1.0
 
  - Added error messages resource to NuGet package (no code changes)